# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CreateRunRequest": {
        "properties": {
          "device": {
            "default": "auto",
            "enum": [
              "auto",
              "cpu",
              "cuda",
              "mps"
            ],
            "title": "Device",
            "type": "string"
          },
          "kpt_thr": {
            "default": 0.3,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Kpt Thr",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "model": {
            "default": "human",
            "enum": [
              "human",
              "wholebody",
              "hand",
              "human3d"
            ],
            "title": "Model",
            "type": "string"
          },
          "session": {
            "title": "Session",
            "type": "string"
          }
        },
        "required": [
          "label",
          "session"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "FrameKeypoints": {
        "description": "One source frame and the artifacts extraction derived from it.",
        "properties": {
          "frame": {
            "title": "Frame",
            "type": "string"
          },
          "keypoints_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keypoints Key"
          },
          "mean_score": {
            "default": 0.0,
            "title": "Mean Score",
            "type": "number"
          },
          "num_instances": {
            "default": 0,
            "title": "Num Instances",
            "type": "integer"
          },
          "num_keypoints": {
            "default": 0,
            "title": "Num Keypoints",
            "type": "integer"
          },
          "overlay_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlay Key"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          }
        },
        "required": [
          "frame",
          "source_key"
        ],
        "title": "FrameKeypoints",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRecord": {
        "description": "The full persisted run manifest.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device": {
            "enum": [
              "auto",
              "cpu",
              "cuda",
              "mps"
            ],
            "title": "Device",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/FrameKeypoints"
            },
            "title": "Frames",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kpt_thr": {
            "title": "Kpt Thr",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "manifest_key": {
            "title": "Manifest Key",
            "type": "string"
          },
          "model": {
            "enum": [
              "human",
              "wholebody",
              "hand",
              "human3d"
            ],
            "title": "Model",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "session": {
            "title": "Session",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "done",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/RunSummary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "session",
          "model",
          "device",
          "kpt_thr",
          "status",
          "created_at",
          "updated_at",
          "manifest_key"
        ],
        "title": "RunRecord",
        "type": "object"
      },
      "RunSummary": {
        "description": "Aggregate metrics for a finished run \u2014 the write-amplification story.",
        "properties": {
          "amplification_ratio": {
            "default": 0.0,
            "title": "Amplification Ratio",
            "type": "number"
          },
          "derived_bytes": {
            "default": 0,
            "title": "Derived Bytes",
            "type": "integer"
          },
          "frame_count": {
            "default": 0,
            "title": "Frame Count",
            "type": "integer"
          },
          "source_bytes": {
            "default": 0,
            "title": "Source Bytes",
            "type": "integer"
          },
          "total_instances": {
            "default": 0,
            "title": "Total Instances",
            "type": "integer"
          },
          "total_keypoints": {
            "default": 0,
            "title": "Total Keypoints",
            "type": "integer"
          }
        },
        "title": "RunSummary",
        "type": "object"
      },
...
      "UpdateRunRequest": {
        "description": "Re-config is a new run, so the source session is intentionally fixed.",
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "UpdateRunRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local API for MMPose Video Keypoint Extraction: ingest video/frame sessions to Backblaze B2, run the on-device MMPose engine to extract 2D/3D skeleton keypoints, and write per-frame keypoint JSON, skeleton-overlay PNGs, and a keypoints_index.jsonl dataset manifest back to B2. This contract documents the local API, not a hosted public endpoint.",
    "title": "MMPose Video Keypoint Extraction API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "operationId": "list_runs_endpoint_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunRecord"
                  },
                  "title": "Response List Runs Endpoint Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs Endpoint",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_endpoint_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run Endpoint",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "delete": {
        "operationId": "delete_run_endpoint_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Run Endpoint",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_endpoint_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Endpoint",
        "tags": [
          "runs"
        ]
      },
      "patch": {
        "operationId": "update_run_endpoint_runs__run_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Run Endpoint",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/execute": {
      "post": {
        "operationId": "execute_run_endpoint_runs__run_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Run Endpoint",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/sessions/{session}/frames": {
      "get": {
        "operationId": "session_frames_endpoint_sessions__session__frames_get",
        "parameters": [
          {
            "in": "path",
            "name": "session",
            "required": true,
            "schema": {
              "title": "Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Session Frames Endpoint Sessions  Session  Frames Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Frames Endpoint",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}